Make record-based query test infrastructure generic
