+++ test/instruction_graph_test_utils.h
@@ -1,269 +1,12 @@
 #pragma once
 
-#include "command_graph_generator_test_utils.h"
+#include "graph_test_utils.h"
 #include "instruction_graph_generator.h"
-
 
 namespace celerity::test_utils {
 
-/// View on a subset of instructions within an instruction_recorder. Allows selecting for subsets by various predicates, sets of successors and predecessor
-/// nodes within the graph. Depending on the selectors used, returned sub-queries are automatically specialized on the record type, which will allow directly
-/// de-referencing a single-element query via the `->` operator.
-template <typename Record = instruction_record>
-class instruction_query {
-  public:
-	using record_type = Record;
-
-	template <typename R = Record, std::enable_if_t<std::is_same_v<R, instruction_record>, int> = 0>
-	explicit instruction_query(const instruction_recorder& recorder)
-	    : instruction_query(&recorder, non_owning_pointers(recorder.get_instructions()), std::string()) {}
-
-	// allow upcast
-	template <typename SpecificRecord, std::enable_if_t<std::is_base_of_v<Record, SpecificRecord> && !std::is_same_v<Record, SpecificRecord>, int> = 0>
-	instruction_query(const instruction_query<SpecificRecord>& other)
-	    : instruction_query(other.m_recorder, std::vector<const Record*>(other.m_result.begin(), other.m_result.end()), other.m_trace) {}
-
-	/// Returns a subset of the selected instructions that matches all provided filters and (optionally) the provided record type.
-	template <typename SpecificRecord = Record, typename... Filters>
-	instruction_query<SpecificRecord> select_all(const Filters&... filters) const {
-		std::vector<const SpecificRecord*> filtered;
-		for(const auto instr : m_result) {
-			if(matches<SpecificRecord>(*instr, filters...)) { filtered.push_back(utils::as<SpecificRecord>(instr)); }
-		}
-		return instruction_query<SpecificRecord>(m_recorder, std::move(filtered),
-		    std::is_same_v<SpecificRecord, Record> && sizeof...(Filters) == 0 ? m_trace : filter_trace<Record, SpecificRecord>("select_all", filters...));
-	}
-
-	/// Like `select_all`, but asserts that the returned query holds exactly one element (and thus can be dereferenced using `->`).
-	template <typename SpecificRecord = Record, typename... Filters>
-	instruction_query<SpecificRecord> select_unique(const Filters&... filters) const {
-		auto query = select_all<SpecificRecord>(filters...).assert_unique();
-		return instruction_query<SpecificRecord>(m_recorder, std::move(query.m_result), filter_trace<Record, SpecificRecord>("select_unique", filters...));
-	}
-
-	/// Selects all predecessors (dependencies) of all nodes in the current query.
-	instruction_query<instruction_record> predecessors() const {
-		std::vector<const instruction_record*> predecessors;
-		// find predecessors without duplicates (even when m_result.size() > 1) and keep them in recorder-ordering
-		for(const auto& maybe_predecessor : m_recorder->get_instructions()) {
-			if(std::any_of(m_recorder->get_dependencies().begin(), m_recorder->get_dependencies().end(), [&](const instruction_dependency_record& dep) {
-				   return dep.predecessor == maybe_predecessor->id
-				          && std::any_of(m_result.begin(), m_result.end(), [&](const Record* me) { return me->id == dep.successor; });
-			   })) {
-				predecessors.push_back(maybe_predecessor.get());
-			}
-		}
-		return instruction_query<instruction_record>(m_recorder, std::move(predecessors), m_trace + ".predecessors()");
-	}
-
-	/// Recursively selects all predecessors of all nodes in the current query.
-	instruction_query<instruction_record> transitive_predecessors() const {
-		for(auto query = predecessors();;) {
-			auto next = union_of(query, query.predecessors());
-			if(query == next) { return instruction_query<instruction_record>(m_recorder, std::move(query.m_result), m_trace + ".transitive_predecessors()"); }
-			query = std::move(next);
-		}
-	};
-
-	/// Recursively selects all predecessors of all nodes in the current query, but only walks across nodes that match all provided filters (and the specific
-	/// record type if provided).
-	template <typename SpecificRecord, typename... Filters>
-	instruction_query<instruction_record> transitive_predecessors_across(const Filters&... filters) const {
-		for(auto query = predecessors();;) {
-			auto next = union_of(query, query.template select_all<SpecificRecord>(filters...).predecessors());
-			if(query == next) {
-				return instruction_query<instruction_record>(
-				    m_recorder, std::move(query.m_result), filter_trace<Record, SpecificRecord>("transitive_predecessors_across", filters...));
-			}
-			query = std::move(next);
-		}
-	};
-
-	/// Selects all successors (dependers) of all nodes in the current query.
-	instruction_query<instruction_record> successors() const {
-		std::vector<const instruction_record*> successors;
-		// find successors without duplicates (even when m_result.size() > 1) and keep them in recorder-ordering
-		for(const auto& maybe_successor : m_recorder->get_instructions()) {
-			if(std::any_of(m_recorder->get_dependencies().begin(), m_recorder->get_dependencies().end(), [&](const instruction_dependency_record& dep) {
-				   return dep.successor == maybe_successor->id
-				          && std::any_of(m_result.begin(), m_result.end(), [&](const Record* me) { return me->id == dep.predecessor; });
-			   })) {
-				successors.push_back(maybe_successor.get());
-			}
-		}
-		return instruction_query<instruction_record>(m_recorder, std::move(successors), m_trace + ".successors()");
-	}
-
-	/// Recursively selects all successors of all nodes in the current query.
-	instruction_query<instruction_record> transitive_successors() const {
-		for(auto query = successors();;) {
-			auto next = union_of(query, query.successors());
-			if(query == next) { return instruction_query<instruction_record>(m_recorder, std::move(query.m_result), m_trace + ".transitive_successors()"); }
-			query = std::move(next);
-		}
-	};
-
-	/// Recursively selects all successors of all nodes in the current query, but only walks across nodes that match all provided filters (and the specific
-	/// record type if provided).
-	template <typename SpecificRecord, typename... Filters>
-	instruction_query<instruction_record> transitive_successors_across(const Filters&... filters) const {
-		for(auto query = successors();;) {
-			auto next = union_of(query, query.template select_all<SpecificRecord>(filters...).successors());
-			if(query == next) {
-				return instruction_query<instruction_record>(
-				    m_recorder, std::move(query.m_result), filter_trace<Record, SpecificRecord>("transitive_successors_across", filters...));
-			}
-			query = std::move(next);
-		}
-	};
-
-	/// Check if `this` and `other` are concurrent, i.e. neither is a transitive predecessor of the other.
-	bool is_concurrent_with(const instruction_query<>& other) const {
-		return !transitive_predecessors().contains(other) && !transitive_successors().contains(other);
-	}
-
-	/// Returns the number of nodes in the query.
-	size_t count() const { return m_result.size(); }
-
-	/// Returns the number of nodes in the query that match all provided filters (and the specified record type, if any)
-	template <typename SpecificRecord = Record, typename... Filters>
-	size_t count(const Filters&... filters) const {
-		return static_cast<size_t>(
-		    std::count_if(m_result.begin(), m_result.end(), [&](const Record* instr) { return matches<SpecificRecord>(*instr, filters...); }));
-	}
-
-	/// Returns sub-query containing exactly the `index`-th node in the current query. Nodes keep the order they were recorded in.
-	instruction_query operator[](const size_t index) const {
-		if(index > m_result.size()) {
-			INFO(fmt::format("query: ", m_trace));
-			INFO(fmt::format("result: {}", *this));
-			FAIL(fmt::format("index {} out of bounds (size: {})", index, m_result.size()));
-		}
-		return instruction_query(m_recorder, {m_result[index]}, fmt::format("{}[{}]", m_trace, index));
-	}
-
-	/// Returns a sequence of single-node queries to iterate over all entries of the current query, in the order the nodes were recorded in.
-	std::vector<instruction_query> iterate() const {
-		std::vector<instruction_query> queries;
-		for(size_t i = 0; i < m_result.size(); ++i) {
-			queries.push_back(instruction_query(m_recorder, {m_result[i]}, fmt::format("{}[{}]", m_trace, i)));
-		}
-		return queries;
-	}
-
-	/// Returns whether all nodes in the current query match all provided filters (and have a specific record type, if provided).
-	template <typename SpecificRecord = Record, typename... Filters>
-	bool all_match(const Filters&... filters) const {
-		std::string non_matching;
-		for(const Record* instr : m_result) {
-			if(!matches<SpecificRecord>(*instr, filters...)) {
-				if(!non_matching.empty()) { non_matching += ", "; }
-				fmt::format_to(std::back_inserter(non_matching), "I{}", instr->id);
-			}
-		}
-		if(non_matching.empty()) return true;
-
-		UNSCOPED_INFO(fmt::format("query: {}", filter_trace<Record, SpecificRecord>("all_match", filters...)));
-		UNSCOPED_INFO(fmt::format("non-matching: {{{}}}", non_matching));
-		return false;
-	}
-
-	/// Asserts that the current query does not `match_all` the provided filters, and casts the query to the specific record type (if provided).
-	template <typename SpecificRecord = Record, typename... Filters>
-	instruction_query<SpecificRecord> assert_all(const Filters&... filters) const {
-		REQUIRE(all_match<SpecificRecord>(filters...));
-		std::vector<const SpecificRecord*> result(m_result.size());
-		std::transform(m_result.begin(), m_result.end(), result.begin(), [](const Record* instr) { return utils::as<SpecificRecord>(instr); });
-		return instruction_query<SpecificRecord>(m_recorder, std::move(result), filter_trace<Record, SpecificRecord>("assert_all", filters...));
-	}
-
-	/// Returns whether `is_concurrent` holds for each pair of single-node sub-queries.
-	bool all_concurrent() const {
-		for(size_t i = 0; i < count(); ++i) {
-			for(size_t j = i + 1; j < count(); ++j) {
-				if(!(*this)[i].is_concurrent_with((*this)[j])) {
-					UNSCOPED_INFO(fmt::format("query: {}", m_trace));
-					UNSCOPED_INFO(fmt::format("result: {}", *this));
-					UNSCOPED_INFO(fmt::format("I{} and I{} are not concurrent", (*this)[i]->id, (*this)[j]->id));
-					return false;
-				}
-			}
-		}
-		return true;
-	}
-
-	/// Returns whether `this` contains all nodes that `subset` also contains.
-	bool contains(const instruction_query& subset) const {
-		return std::all_of(subset.m_result.begin(), subset.m_result.end(), [&](const Record* their) {
-			return std::any_of(m_result.begin(), m_result.end(), [&](const Record* my) { return their->id == my->id; }); //
-		});
-	}
-
-	/// Returns whether this query holds exactly one node and whether that node matches all provided filters.
-	template <typename SpecificRecord = Record, typename... Filters>
-	bool is_unique(const Filters&... filters) const {
-		if(m_result.size() != 1) {
-			UNSCOPED_INFO(fmt::format("query: {}", m_trace));
-			UNSCOPED_INFO(fmt::format("result: {}", *this));
-		}
-		return m_result.size() == 1 && all_match<SpecificRecord>(filters...);
-	}
-
-	/// Asserts that`is_unique` is true for the provided filters, and casts the result to the record type (if specified).
-	template <typename SpecificRecord = Record, typename... Filters>
-	instruction_query<SpecificRecord> assert_unique(const Filters&... filters) const {
-		REQUIRE(is_unique<SpecificRecord>(filters...));
-		return instruction_query<SpecificRecord>(
-		    m_recorder, {utils::as<SpecificRecord>(m_result.front())}, filter_trace<Record, SpecificRecord>("assert_unique", filters...));
-	}
-
-	/// Asserts that this query contains exactly one node, and provides direct access to its record.
-	const Record* operator->() const {
-		REQUIRE(is_unique());
-		return m_result.front();
-	}
-
-	// m_result follows m_recorder ordering, so vector-equality is enough
-	friend bool operator==(const instruction_query& lhs, const instruction_query& rhs) { return lhs.m_result == rhs.m_result; }
-	friend bool operator!=(const instruction_query& lhs, const instruction_query& rhs) { return lhs.m_result != rhs.m_result; }
-
-	/// Returns a query containing the union of nodes from all parameters.
-	template <typename... InstructionQueries>
-	friend instruction_query<> union_of(const instruction_query& head, const InstructionQueries&... tail) {
-		// call through a proper member function, because GCC will not extend friendship with InstructionQueries... to inline-friend functions
-		return head.union_with(tail...);
-	}
-
-	/// Returns a query containing the intersection of nodes from all parameters.
-	template <typename... InstructionQueries>
-	friend instruction_query intersection_of(const instruction_query& head, const InstructionQueries&... tail) {
-		// call through a proper member function, because GCC will not extend friendship with InstructionQueries... to inline-friend functions
-		return head.intersection_with(tail...);
-	}
-
-  private:
-	template <typename>
-	friend class instruction_query;
-
-	template <typename, typename, typename>
-	friend struct fmt::formatter;
-
-	const instruction_recorder* m_recorder;
-	std::vector<const Record*> m_result;
-	std::string m_trace;
-
-	template <typename T>
-	static std::vector<const T*> non_owning_pointers(const std::vector<std::unique_ptr<T>>& unique) {
-		std::vector<const T*> ptrs(unique.size());
-		std::transform(unique.begin(), unique.end(), ptrs.begin(), [](const std::unique_ptr<T>& p) { return p.get(); });
-		return ptrs;
-	}
-
-	template <typename SpecificRecord, typename... Filters>
-	static bool matches(const Record& instr, const Filters&... filters) {
-		return utils::isa<SpecificRecord>(&instr) && (instruction_query<SpecificRecord>::matches(*utils::as<SpecificRecord>(&instr), filters) && ...);
-	}
-
+template <typename Record>
+struct instruction_matcher {
 	static bool matches(const Record& instr, const instruction_id iid) { return instr.id == iid; }
 
 	static bool matches(const Record& instr, const task_id tid) {
@@ -303,63 +46,9 @@ class instruction_query {
 	static std::string print_filter(const Predicate& /* predicate */) {
 		return "<lambda>";
 	}
-
-	template <typename GeneralRecord, typename SpecificRecord, typename... Filters>
-	std::string filter_trace(const std::string& selector, const Filters&... filters) const {
-		auto trace = m_trace.empty() ? selector : fmt::format("{}.{}", m_trace, selector);
-		if constexpr(!std::is_same_v<GeneralRecord, SpecificRecord>) { fmt::format_to(std::back_inserter(trace), "<{}>", kernel_debug_name<SpecificRecord>()); }
-
-		std::string param_trace;
-		// [[maybe_unused]]: GCC will emit an unused-warning if `Filters...` is empty and the comma-fold expression below expands to nothing
-		[[maybe_unused]] const auto format_filter_param = [&](const auto& f) {
-			if(!param_trace.empty()) { param_trace += ", "; }
-			param_trace += instruction_query<SpecificRecord>::print_filter(f);
-		};
-		(format_filter_param(filters), ...);
-		fmt::format_to(std::back_inserter(trace), "({})", param_trace);
-		return trace;
-	}
-
-	template <typename... InstructionQueries>
-	instruction_query<> union_with(const InstructionQueries&... tail) const {
-		assert(((m_recorder == tail.m_recorder) && ...));
-
-		// construct union in recorder-ordering without duplicates - always returns the base type
-		std::vector<const instruction_record*> union_query;
-		for(const auto& instr : m_recorder->get_instructions()) {
-			if(std::find(m_result.begin(), m_result.end(), instr.get()) != m_result.end()
-			    || ((std::find(tail.m_result.begin(), tail.m_result.end(), instr.get()) != tail.m_result.end()) || ...)) { //
-				union_query.push_back(instr.get());
-			}
-		}
-
-		std::string trace = fmt::format("union_of({}", m_trace);
-		(((trace += ", ") += tail.m_trace), ...);
-		trace += ")";
-		return instruction_query<>(m_recorder, std::move(union_query), std::move(trace));
-	}
-
-	template <typename... InstructionQueries>
-	instruction_query intersection_with(const InstructionQueries&... tail) const {
-		assert(((m_recorder == tail.m_recorder) && ...));
-
-		// construct intersection in recorder-ordering without duplicates - returns the type of `head`
-		std::vector<const Record*> intersection_query;
-		for(const auto& instr : m_result) {
-			if(((std::find(tail.m_result.begin(), tail.m_result.end(), instr) != tail.m_result.end()) && ...)) { //
-				intersection_query.push_back(instr);
-			}
-		}
-
-		std::string trace = fmt::format("intersection_of({}", m_trace);
-		(((trace += ", ") += tail.m_trace), ...);
-		trace += ")";
-		return instruction_query(m_recorder, std::move(intersection_query), std::move(trace));
-	}
-
-	instruction_query(const instruction_recorder* recorder, std::vector<const Record*> query, std::string trace)
-	    : m_recorder(recorder), m_result(std::move(query)), m_trace(std::move(trace)) {}
 };
+
+using instruction_query = graph_query<instruction_record, instruction_record, instruction_recorder, instruction_matcher>;
 
 /// A view on an `instruction_recorder` like `instruction_query`, but for `outbound_pilot`s.
 class pilot_query {
@@ -624,7 +313,7 @@ class idag_test_context {
 
 	command_graph_generator& get_graph_generator() { return m_cggen; }
 
-	instruction_query<> query_instructions() const { return instruction_query<>(m_instr_recorder); }
+	instruction_query query_instructions() const { return instruction_query(m_instr_recorder); }
 
 	pilot_query query_outbound_pilots() const { return pilot_query(m_instr_recorder); }
 
@@ -712,22 +401,3 @@ class idag_test_context {
 };
 
 } // namespace celerity::test_utils
-
-template <typename Record>
-struct fmt::formatter<celerity::test_utils::instruction_query<Record>> : fmt::formatter<size_t> {
-	format_context::iterator format(const celerity::test_utils::instruction_query<Record>& iq, format_context& ctx) const {
-		auto out = ctx.out();
-		*out++ = '{';
-		for(size_t i = 0; i < iq.m_result.size(); ++i) {
-			if(i > 0) { out = std::copy_n(", ", 2, out); }
-			fmt::format_to(out, "I{}", iq.m_result[i]->id);
-		}
-		*out++ = '}';
-		return out;
-	}
-};
-
-template <typename Record>
-struct Catch::StringMaker<celerity::test_utils::instruction_query<Record>> {
-	static std::string convert(const celerity::test_utils::instruction_query<Record>& iq) { return fmt::format("{}", iq); }
-};
